+++ database/migrations/price.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Float
 from sqlalchemy.orm import relationship
 
 from database.app import db
@@ -9,7 +9,7 @@ class Price(db.Model):
 
     id = Column(Integer, primary_key=True)
     uc = Column(String)
-    sum = Column(Integer)
+    sum = Column(Float)
     currency = Column(String)
 
     orders = relationship('Orders', back_populates='price')
